requirements.txt
anyio==4.4.0
APScheduler==3.10.4
beautifulsoup4==4.11.1
blinker==1.8.2
cachetools==4.2.2
certifi==2024.8.30
charset-normalizer==2.1.1
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-PyMongo==2.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==4.9.1
MarkupSafe==2.1.5
pymongo==4.8.0
python-dotenv==1.0.0
python-telegram-bot==21.5
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.2.post1
telegram==0.0.1
tornado==6.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
Werkzeug==3.0.4
